updated flow implementation
